Only demonstrate a single good pattern
